@@ -8,7 +8,9 @@ from unique_toolkit._common.chunk_relevancy_sorter.exception import (
 )
 from unique_toolkit._common.chunk_relevancy_sorter.service import ChunkRelevancySorter
 from unique_toolkit.agentic.evaluation.schemas import EvaluationMetricName
+from unique_toolkit.agentic.feature_flags import feature_flags
 from unique_toolkit.agentic.history_manager.utils import transform_chunks_to_string
+from unique_toolkit.agentic.message_log_manager.service import MessageStepLogger
 from unique_toolkit.agentic.tools.agent_chunks_hanlder import AgentChunksHandler
 from unique_toolkit.agentic.tools.factory import ToolFactory
 from unique_toolkit.agentic.tools.schemas import ToolCallResponse
@@ -16,8 +18,10 @@ from unique_toolkit.agentic.tools.tool import Tool
 from unique_toolkit.agentic.tools.tool_progress_reporter import ProgressState
 from unique_toolkit.app.schemas import BaseEvent, ChatEvent, Event
 from unique_toolkit.chat.schemas import (
+    MessageLog,
     MessageLogDetails,
     MessageLogEvent,
+    MessageLogStatus,
 )
 from unique_toolkit.chat.service import LanguageModelToolDescription
 from unique_toolkit.content.schemas import Content, ContentChunk, ContentReference
@@ -50,13 +54,20 @@ class InternalSearchService:
         chunk_relevancy_sorter: ChunkRelevancySorter,
         chat_id: str | None,
         logger: Logger,
+        company_id: str | None = None,
+        message_step_logger: MessageStepLogger | None = None,
+        display_name: str = "Internal Search",
     ):
         self.config = config
         self.content_service = content_service
         self.chunk_relevancy_sorter = chunk_relevancy_sorter
         self.chat_id = chat_id
+        self.company_id = company_id
         self.logger = logger
         self.tool_execution_message_name = "Internal search"
+        self._message_step_logger = message_step_logger
+        self._display_name = display_name
+        self._active_message_log: MessageLog | None = None
 
     async def post_progress_message(self, message: str, *args, **kwargs):
         pass
@@ -161,11 +172,19 @@ class InternalSearchService:
 
         # Apply chunk relevancy sorter if enabled
         if self.config.chunk_relevancy_sort_config.enabled:
-            for i, result in enumerate(found_chunks_per_search_string):
-                await self.post_progress_message(
-                    f"{result.query} (_Resorting {len(result.chunks)} search results_ 🔄 in query {i + 1}/{len(found_chunks_per_search_string)})",
-                    **kwargs,
+            if feature_flags.is_new_answers_ui_enabled(self.company_id):
+                self._active_message_log = (
+                    await self._create_or_update_active_message_log(
+                        progress_message="_Resorting search results_",
+                        search_strings_list=search_strings,
+                    )
                 )
+            for i, result in enumerate(found_chunks_per_search_string):
+                if not feature_flags.is_new_answers_ui_enabled(self.company_id):
+                    await self.post_progress_message(
+                        f"{result.query} (_Resorting {len(result.chunks)} search results_ 🔄 in query {i + 1}/{len(found_chunks_per_search_string)})",
+                        **kwargs,
+                    )
                 result.chunks = await self._resort_found_chunks_if_enabled(
                     found_chunks=result.chunks,
                     search_string=result.query,
@@ -175,17 +194,24 @@ class InternalSearchService:
         # 3. Pick a subset of the search results
         ###
         if (
-            self.config.experimental_features.enable_multiple_search_strings_execution
+            self.config.enable_multiple_search_strings_execution
             and len(found_chunks_per_search_string) > 1
         ):
             found_chunks_per_search_string = interleave_search_results_round_robin(
                 found_chunks_per_search_string
             )
 
-        await self.post_progress_message(
-            f"{', '.join(search_strings)} (_Postprocessing search results_)",
-            **kwargs,
-        )
+        if feature_flags.is_new_answers_ui_enabled(self.company_id):
+            progress_message = "_Postprocessing search results_"
+            self._active_message_log = await self._create_or_update_active_message_log(
+                progress_message=progress_message,
+                search_strings_list=search_strings,
+            )
+        else:
+            await self.post_progress_message(
+                f"{', '.join(search_strings)} (_Postprocessing search results_)",
+                **kwargs,
+            )
         found_chunks = [
             chunk
             for result in found_chunks_per_search_string
@@ -299,6 +325,80 @@ class InternalSearchService:
             )
             return self.config.max_tokens_for_sources
 
+    async def _create_or_update_active_message_log(
+        self,
+        *,
+        progress_message: str | None = None,
+        chunks: list[ContentChunk] | None = None,
+        search_strings_list: list[str],
+        status: MessageLogStatus | None = None,
+    ) -> MessageLog | None:
+        if self._message_step_logger is None:
+            return None
+        message_log_reference_list = []
+        if chunks is not None:
+            message_log_reference_list = (
+                await self._define_reference_list_for_message_log(
+                    content_chunks=chunks,
+                )
+            )
+        details = await self._prepare_message_log_details(
+            query_list=search_strings_list
+        )
+        return self._message_step_logger.create_or_update_message_log(
+            active_message_log=self._active_message_log,
+            header=self._display_name,
+            progress_message=progress_message,
+            details=details,
+            references=message_log_reference_list,
+            **({"status": status} if status is not None else {}),
+        )
+
+    async def _define_reference_list_for_message_log(
+        self,
+        *,
+        content_chunks: list[ContentChunk],
+    ) -> list[ContentReference]:
+        """
+        Create a reference list for internal search content chunks.
+
+        Args:
+            content_chunks: List of ContentChunk objects to convert
+        Returns:
+            List of ContentReference objects
+        """
+        data: list[ContentReference] = []
+        for count, content_chunk in enumerate(content_chunks):
+            reference_name: str = content_chunk.title or content_chunk.key or ""
+
+            data.append(
+                ContentReference(
+                    name=reference_name,
+                    sequence_number=count,
+                    source="internal",
+                    source_id=content_chunk.id,
+                    url=f"unique://content/{content_chunk.id}",
+                )
+            )
+            count += 1
+
+        return data
+
+    async def _prepare_message_log_details(
+        self, *, query_list: list[str]
+    ) -> MessageLogDetails:
+        details = MessageLogDetails(
+            data=[
+                MessageLogEvent(
+                    type="InternalSearch",
+                    text=query_for_log,
+                )
+                for query_for_log in query_list
+            ]
+        )
+
+        return details
+
 
 class InternalSearchTool(Tool[InternalSearchConfig], InternalSearchService):
     name = "InternalSearch"
@@ -319,13 +419,17 @@ class InternalSearchTool(Tool[InternalSearchConfig], InternalSearchService):
             chat_id = self.event.payload.chat_id
         else:
             chat_id = None
+        self._display_name = kwargs.get("display_name", "Internal Search")
         InternalSearchService.__init__(
             self,
             config=configuration,
             content_service=content_service,
             chunk_relevancy_sorter=chunk_relevancy_sorter,
             chat_id=chat_id,
+            company_id=self.event.company_id,
             logger=self.logger,
+            message_step_logger=self._message_step_logger,
+            display_name=self._display_name,
         )
 
     async def post_progress_message(
@@ -355,7 +459,7 @@ class InternalSearchTool(Tool[InternalSearchConfig], InternalSearchService):
     @override
     def tool_description(self) -> LanguageModelToolDescription:
         # Conditionally set the type based on config
-        if self.config.experimental_features.enable_multiple_search_strings_execution:
+        if self.config.enable_multiple_search_strings_execution:
             search_string_field = (
                 list[str],
                 Field(
@@ -438,16 +542,26 @@ class InternalSearchTool(Tool[InternalSearchConfig], InternalSearchService):
         search_strings_list = list(dict.fromkeys(search_strings_list))
         search_strings_list = search_strings_list[: self.config.max_search_strings]
 
-        await self.post_progress_message(f"{'; '.join(search_strings_list)}", tool_call)
+        self._active_message_log = await self._create_or_update_active_message_log(
+            progress_message="Retrieving search results...",
+            search_strings_list=search_strings_list,
+        )
+
+        if not feature_flags.is_new_answers_ui_enabled(self.company_id):
+            await self.post_progress_message(
+                f"{'; '.join(search_strings_list)}", tool_call
+            )
 
         selected_chunks = await self.search(
             **tool_call.arguments,
             tool_call=tool_call,  # Need to pass tool_call to post_progress_message
+            active_message_log=self._active_message_log,
         )
 
-        ## Message log entry
-        await self._create_message_log_entry(
-            chunks=selected_chunks, search_strings_list=search_strings_list
+        self._active_message_log = await self._create_or_update_active_message_log(
+            chunks=selected_chunks,
+            search_strings_list=search_strings_list,
+            status=MessageLogStatus.COMPLETED,
         )
 
         ## Modify metadata in chunks
@@ -463,7 +577,9 @@ class InternalSearchTool(Tool[InternalSearchConfig], InternalSearchService):
             debug_info=self.debug_info,
         )
 
-        if self.tool_progress_reporter:
+        if self.tool_progress_reporter and not feature_flags.is_new_answers_ui_enabled(
+            self.company_id
+        ):
             await self.tool_progress_reporter.notify_from_tool_call(
                 tool_call=tool_call,
                 name=f"**{self.tool_execution_message_name}**",
@@ -472,69 +588,6 @@ class InternalSearchTool(Tool[InternalSearchConfig], InternalSearchService):
             )
 
         return tool_response
-
-    async def _create_message_log_entry(
-        self, *, chunks: list[ContentChunk], search_strings_list: list[str]
-    ) -> None:
-        message_log_reference_list: list[
-            ContentReference
-        ] = await self._define_reference_list_for_message_log(
-            content_chunks=chunks,
-        )
-        details = await self._prepare_message_log_details(
-            query_list=search_strings_list
-        )
-        self._message_step_logger.create_message_log_entry(
-            text=f"**{self.display_name()}**",
-            details=details,
-            references=message_log_reference_list,
-        )
-
-    async def _define_reference_list_for_message_log(
-        self,
-        *,
-        content_chunks: list[ContentChunk],
-    ) -> list[ContentReference]:
-        """
-        Create a reference list for internal search content chunks.
-
-        Args:
-            content_chunks: List of ContentChunk objects to convert
-        Returns:
-            List of ContentReference objects
-        """
-        data: list[ContentReference] = []
-        count = 0
-        for content_chunk in content_chunks:
-            reference_name: str = content_chunk.title or content_chunk.key or ""
-
-            data.append(
-                ContentReference(
-                    name=reference_name,
-                    sequence_number=count,
-                    source="internal",
-                    source_id=content_chunk.id,
-                    url=f"unique://content/{content_chunk.id}",
-                )
-            )
-            count += 1
-
-        return data
-
-    async def _prepare_message_log_details(
-        self, *, query_list: list[str]
-    ) -> MessageLogDetails:
-        details = MessageLogDetails(
-            data=[
-                MessageLogEvent(
-                    type="InternalSearch",
-                    text=query_for_log,
-                )
-                for query_for_log in query_list
-            ]
-        )
-
-        return details
 
     ## Note: This function is only used by the Investment Research Agent and Agentic Search. Once these agents are moved out of the monorepo, this function should be removed.
     def get_tool_call_result_for_loop_history(
