@@ -1,13 +1,16 @@
+from typing import Annotated
+
 from pydantic import Field
 from pydantic.json_schema import SkipJsonSchema
 from unique_toolkit import LanguageModelName
+from unique_toolkit._common.pydantic.rjsf_tags import RJSFMetaTag
 from unique_toolkit._common.validators import LMI, get_LMI_default_field
 from unique_toolkit.agentic.tools.schemas import BaseToolConfig
 
 from unique_swot.services.generation.config import ReportGenerationConfig
-from unique_swot.services.generation.extraction.config import ExtractionConfig
 from unique_swot.services.report import ReportRendererConfig
 from unique_swot.services.source_management.config import SourceManagementConfig
+from unique_swot.services.summarization.config import SummarizationConfig
 
 TOOL_DESCRIPTION = """
 This tool is used to perfom a SWOT analysis of a company by analyzing its strengths, weaknesses, opportunities, and threats.
@@ -15,7 +18,7 @@ The user can either generate a new SWOT analysis or modify an existing one.
 If the user simply says RUN, the tool will generate a new SWOT analysis.
 
 If the user simply says RUN, It means that he expects the tool to generate a new SWOT analysis.
-"""
+""".strip()
 
 
 class SwotAnalysisToolConfig(BaseToolConfig):
@@ -31,10 +34,6 @@ class SwotAnalysisToolConfig(BaseToolConfig):
         default_factory=SourceManagementConfig,
         description="The configuration for the source management.",
     )
-    extraction_config: ExtractionConfig = Field(
-        default_factory=ExtractionConfig,
-        description="The configuration for the extraction.",
-    )
     report_generation_config: ReportGenerationConfig = Field(
         default_factory=ReportGenerationConfig,
         description="The configuration for the report generation.",
@@ -43,11 +42,21 @@ class SwotAnalysisToolConfig(BaseToolConfig):
         default_factory=ReportRendererConfig,
         description="The configuration for the report renderer.",
     )
-    tool_description: str = Field(
+    report_summarization_config: SummarizationConfig = Field(
+        default_factory=SummarizationConfig,
+        description="The configuration for the summarization.",
+    )
+    tool_description: Annotated[
+        str,
+        RJSFMetaTag.StringWidget.textarea(rows=len(TOOL_DESCRIPTION.split("\n"))),
+    ] = Field(
         default=TOOL_DESCRIPTION,
         description="The description of the SWOT analysis tool.",
     )
-    tool_description_for_system_prompt: str = Field(
+    tool_description_for_system_prompt: Annotated[
+        str,
+        RJSFMetaTag.StringWidget.textarea(rows=len(TOOL_DESCRIPTION.split("\n"))),
+    ] = Field(
         default=TOOL_DESCRIPTION,
         description="The system prompt for the SWOT analysis tool.",
     )
