@@ -7,10 +7,12 @@ from unique_toolkit.language_model.infos import (
     LanguageModelName,
 )
 
-from unique_swot.services.generation.extraction.config import (
-    ExtractionConfig,
-)
-from unique_swot.services.generation.reporting.config import ReportingConfig
+from unique_swot.services.generation.agentic.config import AgenticGeneratorConfig
+
+# from unique_swot.services.generation.extraction.config import (
+#     ExtractionConfig,
+# )
+# from unique_swot.services.generation.reporting.config import ReportingConfig
 
 _LOGGER = getLogger(__name__)
 
@@ -36,12 +38,7 @@ class ReportGenerationConfig(BaseModel):
     language_model: LMI = get_LMI_default_field(
         _DEFAULT_LANGUAGE_MODEL, description="The language model to use for generation"
     )
-    extraction_config: ExtractionConfig = Field(
-        default_factory=ExtractionConfig,
-        description="The configuration for the extraction.",
-    )
-
-    reporting_config: ReportingConfig = Field(
-        default_factory=ReportingConfig,
-        description="The configuration for the reporting.",
+    agentic_generator_config: AgenticGeneratorConfig = Field(
+        default_factory=AgenticGeneratorConfig,
+        description="The configuration for the agentic generator.",
     )
