@@ -1,192 +1,136 @@
+from logging import getLogger
 from typing import AsyncIterator, Protocol
 
 from tqdm.asyncio import tqdm
-from unique_toolkit.content.schemas import ContentReference
+from unique_toolkit.content import Content, ContentChunk, ContentReference
 
-from unique_swot.services.generation.context import SWOTComponent
-from unique_swot.services.generation.extraction.models import SWOTExtractionModel
-from unique_swot.services.generation.reporting.models import SWOTConsolidatedReport
+from unique_swot.services.generation.models.base import SWOTReportComponents
 from unique_swot.services.memory.base import SwotMemoryService
-from unique_swot.services.schemas import SWOTOperation, SWOTPlan
-from unique_swot.services.source_management.schema import Source
+from unique_swot.services.schemas import SWOTPlan
 from unique_swot.services.source_management.selection.schema import (
     SourceSelectionResult,
 )
 
+_LOGGER = getLogger(__name__)
 
-class Notifier(Protocol):
-    """This class is responsible for notifying the user of the progress of the SWOT analysis."""
 
-    def set_progress_total_steps(self, total_steps: int): ...
+class StepNotifier(Protocol):
+    """This class is responsible for notifying the user of the progress of a specific step of the SWOT analysis."""
 
     async def notify(
-        self, title: str, description: str = "", sources: list[ContentReference] = []
-    ): ...
-
-    async def init_progress(self, session_info: str): ...
-
-    async def increment_progress(self, step_increment: float, progress_info: str): ...
-
-    async def end_progress(
-        self, failed: bool = False, failure_message: str | None = None
+        self,
+        title: str,
+        description: str = "",
+        sources: list[ContentReference] = [],
+        progress: int | None = None,
+        completed: bool = False,
     ): ...
 
 
 class SourceCollector(Protocol):
     """This class is responsible for collecting the sources from the various data sources."""
 
-    async def collect(self) -> list[Source]: ...
+    async def collect(self, *, step_notifier: StepNotifier) -> list[Content]: ...
 
 
 class SourceSelector(Protocol):
     """This class is responsible for selecting the sources that are most relevant to the SWOT analysis."""
 
     async def select(
-        self, *, company_name: str, source: Source
+        self, *, company_name: str, content: Content, step_notifier: StepNotifier
     ) -> SourceSelectionResult: ...
 
 
 class SourceIterator(Protocol):
     """This class is responsible for prioritizing the sources that are most relevant to the SWOT analysis."""
 
-    async def iterate(self, *, sources: list[Source]) -> AsyncIterator[Source]: ...
+    async def iterate(
+        self, *, contents: list[Content], step_notifier: StepNotifier
+    ) -> AsyncIterator[Content]: ...
 
 
-class Extractor(Protocol):
-    """This class is responsible for extracting the SWOT analysis from the sources."""
+class SourceRegistry(Protocol):
+    """This class is responsible for registering the sources that are processed."""
 
-    async def extract(
+    def register(self, *, chunk: ContentChunk) -> str: ...
+
+    def retrieve(self, *, id: str) -> ContentChunk | None: ...
+
+
+class ReportingAgent(Protocol):
+    """This class is responsible for generating the SWOT report."""
+
+    async def generate(
         self,
         *,
-        company_name: str,
-        component: SWOTComponent,
-        source: Source,
-        optional_instruction: str | None,
-    ) -> SWOTExtractionModel: ...
+        plan: SWOTPlan,
+        content: Content,
+        source_registry: SourceRegistry,
+        step_notifier: StepNotifier,
+    ) -> None: ...
 
-
-class ReportManager(Protocol):
-    """This class is responsible for loading the SWOT analysis from the sources."""
-
-    async def generate_and_update_memory(
-        self,
-        *,
-        company_name: str,
-        component: SWOTComponent,
-        extraction_result: SWOTExtractionModel,
-        optional_instruction: str | None,
-    ) -> SWOTConsolidatedReport: ...
-
-    def get_report(self) -> list[SWOTConsolidatedReport]: ...
+    def get_reports(self) -> SWOTReportComponents: ...
 
 
 class SWOTOrchestrator:
     def __init__(
         self,
-        notifier: Notifier,
+        step_notifier: StepNotifier,
         source_collector: SourceCollector,
         source_selector: SourceSelector,
         source_iterator: SourceIterator,
-        extractor: Extractor,
-        report_manager: ReportManager,
+        reporting_agent: ReportingAgent,
+        source_registry: SourceRegistry,
         memory_service: SwotMemoryService,
     ):
         self._source_collector = source_collector
         self._source_selector = source_selector
-        self._extractor = extractor
-        self._report_manager = report_manager
+        self._reporting_agent = reporting_agent
         self._source_iterator = source_iterator
         self._memory_service = memory_service
-        self._notifier = notifier
+        self._step_notifier = step_notifier
+        self._source_registry = source_registry
 
-    async def run(
-        self, *, company_name: str, plan: SWOTPlan
-    ) -> list[SWOTConsolidatedReport]:
-        sources = await self._source_collector.collect()
-
-        await self._notifier.notify(title="Sorting sources by date")
-        await self._notifier.increment_progress(
-            step_increment=0, progress_info="Sorting sources by date"
+    async def run(self, *, company_name: str, plan: SWOTPlan) -> SWOTReportComponents:
+        contents = await self._source_collector.collect(
+            step_notifier=self._step_notifier
         )
 
-        source_iterator = await self._source_iterator.iterate(sources=sources)
+        source_iterator = await self._source_iterator.iterate(
+            contents=contents, step_notifier=self._step_notifier
+        )
 
-        total_steps = len(sources)
-        self._notifier.set_progress_total_steps(total_steps=total_steps)
+        total_steps = len(contents)
 
-        async for source in tqdm(
+        async for content in tqdm(
             source_iterator, total=total_steps, desc="Processing sources"
         ):
-            title = f"Processing source `{source.title}`"
-
-            await self._notifier.notify(title=title)
-
             source_selection_result = await self._source_selector.select(
-                company_name=company_name, source=source
-            )
-            await self._notifier.notify(
-                title=title,
-                description=source_selection_result.notification_message,
-            )
-            await self._notifier.increment_progress(
-                step_increment=0,
-                progress_info=source_selection_result.progress_notification_message,
+                company_name=company_name,
+                content=content,
+                step_notifier=self._step_notifier,
             )
 
             if not source_selection_result.should_select:
                 # Skip the source if it is not selected
-                await self._notifier.increment_progress(
-                    step_increment=1,
-                    progress_info=title,
+                _LOGGER.info(
+                    f"Skipping source `{_get_content_title(content)}` as it is not selected"
                 )
                 continue
-
-            for component in SWOTComponent:
-                step = plan.get_step_result(component)
-                if step.operation == SWOTOperation.NOT_REQUESTED:
-                    await self._notifier.increment_progress(
-                        step_increment=0.25,
-                        progress_info=f"Skipping **{component.value}** as it was not requested...",
-                    )
-                    continue
-
-                await self._notifier.notify(
-                    title=title,
-                    description=f"Processing **{component.value}**...",
+            else:
+                _LOGGER.info(
+                    f"Selecting source `{_get_content_title(content)}` as it is selected"
                 )
 
-                extraction_result = await self._extractor.extract(
-                    company_name=company_name,
-                    component=component,
-                    source=source,
-                    optional_instruction=step.modify_instruction,
-                )
+            await self._reporting_agent.generate(
+                plan=plan,
+                content=content,
+                step_notifier=self._step_notifier,
+                source_registry=self._source_registry,
+            )
 
-                await self._notifier.notify(
-                    title=title,
-                    description=extraction_result.notification_message,
-                )
-                await self._notifier.increment_progress(
-                    step_increment=0.125,
-                    progress_info=extraction_result.progress_notification_message,
-                )
+        return self._reporting_agent.get_reports()
 
-                memory_update_result = (
-                    await self._report_manager.generate_and_update_memory(
-                        company_name=company_name,
-                        component=component,
-                        extraction_result=extraction_result,
-                        optional_instruction=step.modify_instruction,
-                    )
-                )
 
-                await self._notifier.notify(
-                    title=title,
-                    description=memory_update_result.notification_message,
-                )
-                await self._notifier.increment_progress(
-                    step_increment=0.125,
-                    progress_info=memory_update_result.progress_notification_message,
-                )
-
-        return self._report_manager.get_report()
+def _get_content_title(content: Content) -> str:
+    return content.title or content.key or "Unknown Title"
