@@ -1,8 +1,10 @@
 from enum import StrEnum
 from pathlib import Path
+from typing import Annotated
 
 from pydantic import BaseModel, Field
 from unique_toolkit._common.docx_generator import DocxGeneratorConfig
+from unique_toolkit._common.pydantic.rjsf_tags import RJSFMetaTag
 from unique_toolkit._common.pydantic_helpers import get_configuration_dict
 
 from unique_swot.utils import load_template
@@ -17,12 +19,16 @@ REPORT_TEMPLATE: str = load_template(_PROMPTS_DIR, "report_template.j2")
 class DocxRendererType(StrEnum):
     DOCX = "docx"
     CHAT = "chat"
+    STREAM = "stream"
 
 
 class ReportRendererConfig(BaseModel):
     model_config = get_configuration_dict()
 
-    report_template: str = Field(
+    report_template: Annotated[
+        str,
+        RJSFMetaTag.StringWidget.textarea(rows=len(REPORT_TEMPLATE.split("\n"))),
+    ] = Field(
         default=REPORT_TEMPLATE,
         description="Jinja2 template for the report.",
     )
