@@ -2,8 +2,8 @@ from pydantic import BaseModel, Field
 from unique_toolkit._common.pydantic_helpers import get_configuration_dict
 
 from unique_swot.services.source_management.collection.config import EarningsCallConfig
-from unique_swot.services.source_management.iteration.date_relevancy import (
-    DateRelevancySourceIteratorConfig,
+from unique_swot.services.source_management.iteration.config import (
+    SourceIterationConfig,
 )
 from unique_swot.services.source_management.selection.config import (
     SourceSelectionConfig,
@@ -23,7 +23,7 @@ class SourceManagementConfig(BaseModel):
         description="The configuration for the earnings call management.",
     )
 
-    date_relevancy_config: DateRelevancySourceIteratorConfig = Field(
-        default_factory=DateRelevancySourceIteratorConfig,
+    source_iteration_config: SourceIterationConfig = Field(
+        default_factory=SourceIterationConfig,
         description="The configuration for the date relevancy.",
     )
