@@ -1,10 +1,18 @@
-from fastmcp import Client
 import asyncio
+import json
+
+from fastmcp import Client
+
 
 async def main():
-    async with Client("<YOUR_MCP_SERVER_URL>/mcp", auth="oauth") as client:
-
+    async with Client("http://localhost:8003/mcp", auth="oauth") as client:
         print("✓ Authenticated!")
-        
+        tools = await client.list_tools()
+
+        for tool in tools:
+            print("TOOL: ", tool.name)
+            print("TOOL SCHEMA: ", json.dumps(tool.model_dump(), indent=4))
+
+
 if __name__ == "__main__":
     asyncio.run(main())
