@@ -8,6 +8,14 @@ async def test_history_updated_before_reference_extraction(monkeypatch):
     # Lazy import to avoid heavy dependencies at module import time
     from unique_orchestrator.unique_ai import UniqueAI
 
+    # Mock feature_flags to avoid calling the real implementation
+    mock_feature_flags = MagicMock()
+    mock_feature_flags.is_new_answers_ui_enabled = MagicMock(return_value=True)
+    monkeypatch.setattr(
+        "unique_orchestrator.unique_ai.feature_flags",
+        mock_feature_flags,
+    )
+
     # Create a minimal UniqueAI instance with mocked dependencies
     mock_logger = MagicMock()
 
@@ -24,6 +32,7 @@ async def test_history_updated_before_reference_extraction(monkeypatch):
 
     dummy_event = MagicMock()
     dummy_event.payload = DummyEvent.Payload()
+    dummy_event.company_id = "test_company_id"
 
     mock_config = MagicMock()
     mock_config.agent.max_loop_iterations = 1
@@ -33,7 +42,6 @@ async def test_history_updated_before_reference_extraction(monkeypatch):
 
     # Managers
     mock_history_manager = MagicMock()
-    mock_history_manager.has_no_loop_messages.return_value = True
     mock_history_manager._append_tool_calls_to_history = MagicMock()
     mock_history_manager.add_tool_call_results = MagicMock()
 
@@ -79,6 +87,7 @@ async def test_history_updated_before_reference_extraction(monkeypatch):
         return_value=DummyStreamResponse()
     )
     mock_message_step_logger = MagicMock()
+    mock_loop_iteration_runner = AsyncMock(return_value=DummyStreamResponse())
 
     # Instantiate
     ua = UniqueAI(
@@ -97,6 +106,7 @@ async def test_history_updated_before_reference_extraction(monkeypatch):
         postprocessor_manager=MagicMock(),
         message_step_logger=mock_message_step_logger,
         mcp_servers=[],
+        loop_iteration_runner=mock_loop_iteration_runner,
     )
 
     # Bypass Jinja template compilation by stubbing prompt renderers
