@@ -1,25 +1,35 @@
+from unique_toolkit.agentic.loop_runner._iteration_handler_utils import (
+    handle_forced_tools_iteration,
+    handle_last_iteration,
+    handle_normal_iteration,
+)
 from unique_toolkit.agentic.loop_runner.base import LoopIterationRunner
 from unique_toolkit.agentic.loop_runner.middleware import (
-    QWEN_FORCED_TOOL_CALL_PROMPT_INSTRUCTION,
     PlanningConfig,
     PlanningMiddleware,
     PlanningSchemaConfig,
-    QwenForcedToolCallMiddleware,
-    is_qwen_model,
 )
 from unique_toolkit.agentic.loop_runner.runners import (
+    QWEN_FORCED_TOOL_CALL_INSTRUCTION,
+    QWEN_LAST_ITERATION_INSTRUCTION,
     BasicLoopIterationRunner,
     BasicLoopIterationRunnerConfig,
+    QwenLoopIterationRunner,
+    is_qwen_model,
 )
 
 __all__ = [
     "LoopIterationRunner",
-    "QWEN_FORCED_TOOL_CALL_PROMPT_INSTRUCTION",
     "PlanningConfig",
     "PlanningMiddleware",
     "PlanningSchemaConfig",
-    "QwenForcedToolCallMiddleware",
+    "QwenLoopIterationRunner",
     "is_qwen_model",
     "BasicLoopIterationRunnerConfig",
     "BasicLoopIterationRunner",
+    "handle_forced_tools_iteration",
+    "handle_last_iteration",
+    "handle_normal_iteration",
+    "QWEN_FORCED_TOOL_CALL_INSTRUCTION",
+    "QWEN_LAST_ITERATION_INSTRUCTION",
 ]
