@@ -2,5 +2,18 @@ from unique_toolkit.agentic.loop_runner.runners.basic import (
     BasicLoopIterationRunner,
     BasicLoopIterationRunnerConfig,
 )
+from unique_toolkit.agentic.loop_runner.runners.qwen import (
+    QWEN_FORCED_TOOL_CALL_INSTRUCTION,
+    QWEN_LAST_ITERATION_INSTRUCTION,
+    QwenLoopIterationRunner,
+    is_qwen_model,
+)
 
-__all__ = ["BasicLoopIterationRunnerConfig", "BasicLoopIterationRunner"]
+__all__ = [
+    "BasicLoopIterationRunnerConfig",
+    "BasicLoopIterationRunner",
+    "QwenLoopIterationRunner",
+    "QWEN_FORCED_TOOL_CALL_INSTRUCTION",
+    "QWEN_LAST_ITERATION_INSTRUCTION",
+    "is_qwen_model",
+]
