@@ -8,6 +8,7 @@ from collections import defaultdict
 from logging import getLogger
 
 from unique_toolkit.chat.schemas import (
+    MessageLog,
     MessageLogDetails,
     MessageLogStatus,
     MessageLogUncitedReferences,
@@ -61,29 +62,23 @@ class MessageStepLogger:
         self,
         *,
         text: str,
+        status: MessageLogStatus = MessageLogStatus.COMPLETED,
         details: MessageLogDetails | None = None,
-        references: list[ContentReference],
-    ) -> None:
+        references: list[ContentReference] = [],
+    ) -> MessageLog | None:
         """
         Create a full message log entry with question, hits, and references.
-
-        Args:
-            text: The prepared string for the message log entry
-            details: Some formal details about the message log entry
-            references: List of ContentReference objects to reference
+        The created message log entry is returned. If the message log entry is not created, None is returned.
         """
-
-        # Creating a new message log entry with the found hits.
         if not self._chat_service._assistant_message_id:
             _LOGGER.warning(
                 "Assistant message id is not set. Skipping message log entry creation."
             )
             return
-
-        _ = self._chat_service.create_message_log(
+        return self._chat_service.create_message_log(
             message_id=self._chat_service._assistant_message_id,
             text=text,
-            status=MessageLogStatus.COMPLETED,
+            status=status,
             order=self._get_next_message_order(
                 message_id=self._chat_service._assistant_message_id
             ),
@@ -91,3 +86,84 @@ class MessageStepLogger:
             uncited_references=MessageLogUncitedReferences(data=references),
             references=[],
         )
+
+    def update_message_log_entry(
+        self,
+        *,
+        message_log: MessageLog,
+        text: str | None = None,
+        status: MessageLogStatus,
+        details: MessageLogDetails | None = None,
+        references: list[ContentReference] = [],
+    ) -> MessageLog:
+        """
+        Update a message log entry with a new status.
+        The updated message log entry is returned. If the message log entry is not updated, None is returned.
+        """
+        if message_log.message_log_id is None:
+            _LOGGER.warning(
+                "Message log id is not set. Skipping message log entry update."
+            )
+            return message_log
+        return self._chat_service.update_message_log(
+            message_log_id=message_log.message_log_id,
+            order=message_log.order,
+            text=text,
+            status=status,
+            details=details or MessageLogDetails(data=[]),
+            uncited_references=MessageLogUncitedReferences(data=references),
+            references=[],
+        )
+
+    def create_or_update_message_log(
+        self,
+        *,
+        active_message_log: MessageLog | None,
+        header: str,
+        progress_message: str | None = None,
+        status: MessageLogStatus = MessageLogStatus.RUNNING,
+        details: MessageLogDetails | None = None,
+        references: list[ContentReference] | None = [],
+    ) -> MessageLog | None:
+        """
+        Create a new message log entry or update an existing one.
+
+        This is a convenience method that handles the common pattern of:
+        - Creating a new message log if active_message_log is None
+        - Updating the existing message log otherwise
+
+        Args:
+            active_message_log: The current active message log, or None if none exists
+            header: The header to show in bold (e.g., "Internal Search", "Web Search")
+            progress_message: Optional progress message to append after the display name
+            status: The status of the message log
+            details: Optional message log details
+            references: Optional list of content references
+
+        Returns:
+            The created or updated MessageLog, or None if the operation failed
+        """
+        text = (
+            f"**{header}**\n{progress_message}"
+            if progress_message is not None
+            else f"**{header}**"
+        )
+
+        if references is None:
+            references = []
+
+        if active_message_log is None:
+            return self.create_message_log_entry(
+                text=text,
+                details=details,
+                references=references,
+                status=status,
+            )
+        else:
+            return self.update_message_log_entry(
+                message_log=active_message_log,
+                text=text,
+                status=status,
+                details=details,
+                references=references,
+            )
