@@ -1,7 +1,8 @@
 import json
+from typing import Any
 
 from humps import camelize
-from pydantic import BaseModel, ConfigDict, Field, field_validator
+from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
 
 model_config = ConfigDict(
     alias_generator=camelize,
@@ -17,27 +18,31 @@ class ShortTermMemory(BaseModel):
     key: str = Field(alias="object")
     chat_id: str | None
     message_id: str | None
-    data: str | dict | int | float | bool | list | None
+    data: str | dict | int | float | bool | list | None = Field(deprecated=True)
+    value: str | dict[str, Any] = Field(default_factory=dict)
 
-    @field_validator("chat_id", "message_id", mode="before")
-    def validate_chat_id_or_message_id(cls, v, info):
-        field_name = info.field_name
-        data = info.data
+    @model_validator(mode="after")
+    def _data_to_value(self) -> "ShortTermMemory":
+        if isinstance(self.data, dict):
+            self.value = self.data
+        elif isinstance(self.data, str):
+            try:
+                self.value = json.loads(self.data)
+            except json.JSONDecodeError:
+                self.value = self.data
+        elif self.data is None:
+            self.value = ""
+        else:
+            self.value = str(self.data)
+        return self
 
-        # Get the other field's value
-        other_field = "message_id" if field_name == "chat_id" else "chat_id"
-        other_value = data.get(other_field)
-
-        # Check if both are None
-        if v is None and other_value is None:
-            camel_name = camelize(field_name)
-            raise ValueError(
-                f"Either {camel_name} or messageId must be provided"
-                if field_name == "chat_id"
-                else f"Either chatId or {camel_name} must be provided"
-            )
-
-        return v
+    @model_validator(mode="after")
+    def validate_message_id_and_chat_id(self):
+        if (self.message_id is None and self.chat_id is None) or (
+            self.message_id is not None and self.chat_id is not None
+        ):
+            raise ValueError("Either message_id or chat_id must be provided")
+        return self
 
     @field_validator("data", mode="before")
     def validate_data(cls, v):
